submodule deinit: unset core.worktree
When a submodule is deinit'd, the working tree is gone, so the setting of
core.worktree is bogus. Unset it. As we covered the only other case in
which a submodule loses its working tree in the earlier step
(i.e. switching branches of top-level project to move to a commit that did
not have the submodule), this makes the code always maintain
core.worktree correctly unset when there is no working tree
for a submodule.

This re-introduces bc2248f3ce (submodule deinit: unset core.worktree,
2018-06-18), which was reverted as part of e643ae80df (Revert "Merge
branch 'sb/submodule-core-worktree'", 2018-09-07)

The whole series was reverted as the offending commit d9312b830c
(submodule: ensure core.worktree is set after update, 2018-06-18)
was relied on by other commits such as bc2248f3ce.

Keep the offending commit reverted, but its functionality came back via
8b2b64ac7f (Merge branch 'sb/submodule-update-in-c', 2018-09-17), such
that we can reintroduce bc2248f3ce now.

Signed-off-by: Stefan Beller <sbeller@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>